Changed long to double in product

+++ src/main/java/com/inso/model/Productos.java
@@ -23,7 +23,6 @@ import javax.validation.constraints.NotNull;
  */
 @Entity
 @Table(name = "productos")
-
 @NamedQueries({
     @NamedQuery(name = "Productos.findByCIFNombre", query = "SELECT p FROM Productos p WHERE p.productosPK.cIFFarmacia = :cif AND p.productosPK.nombre = :nombre"),
     @NamedQuery(name = "Productos.findByNombre", query = "SELECT p FROM Productos p WHERE p.productosPK.nombre = :nombre"),
@@ -36,7 +35,7 @@ public class Productos implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "Precio")
-    private long precio;
+    private double precio;
     @Basic(optional = false)
     @NotNull
     @Column(name = "Cuantia")
@@ -52,7 +51,7 @@ public class Productos implements Serializable {
         this.productosPK = productosPK;
     }
 
-    public Productos(ProductosPK productosPK, long precio, int cuantia) {
+    public Productos(ProductosPK productosPK, double precio, int cuantia) {
         this.productosPK = productosPK;
         this.precio = precio;
         this.cuantia = cuantia;
@@ -70,11 +69,11 @@ public class Productos implements Serializable {
         this.productosPK = productosPK;
     }
 
-    public long getPrecio() {
+    public double getPrecio() {
         return precio;
     }
 
-    public void setPrecio(long precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
